requirements.txt
opencv-python>=4.8.0
pytesseract>=0.3.10
python-dotenv>=1.0.0
openai>=1.0.0
numpy>=1.24.0 